minnor fix

+++ week5/Main.java
@@ -1,11 +1,7 @@
 import java.sql.*;
 
 public class Main {
-    private static Connection connection;
-
     public static void main(String[] args) {
-        connectToDB();
-
         ResultSet resultSet = selectAllFrom("cars");
         if(resultSet != null){
             try {
@@ -29,26 +25,20 @@ public class Main {
         }
 
         addNewCar("brand1", "model", "bodyStyle", "2014-01-28", "color", 123D);
-        deleteCarById(1);
+        deleteCarById(10);
         updateCarsBrand("newBrand1", 5);
-
-        try {
-            connection.close();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-
 
     }
 
     /***
-     * Подключается к базе данных и
-     * инициирует статическую переменную класса "connection"
+     * Подключается к базе данных 
      */
-    private static void connectToDB(){
+    private static Connection getConnection(){
         final String databaseURL = "jdbc:mysql://localhost:3306/week4";
         final String userName = "root";
         final String password = "pa$$word";
+
+        Connection connection = null;
 
         try {
             Class.forName("com.mysql.cj.jdbc.Driver");// com.mysql.cj.jdbc.Driver   com.mysql.jdbc.Driver
@@ -58,6 +48,7 @@ public class Main {
         } catch (SQLException e) {
             e.printStackTrace();
         }
+        return connection;
     }
 
     /***
@@ -66,17 +57,27 @@ public class Main {
      * @param id - id записи.
      */
     private static void updateCarsBrand(String brand, long id){
-        try {
-            PreparedStatement preparedStatement = connection.prepareStatement("UPDATE cars SET brand = ? WHERE id = ?");
-            preparedStatement.setString(1, brand);
-            preparedStatement.setDouble(2, id);
+        Connection connection = getConnection();
+        if(connection != null) {
+            try {
+                PreparedStatement preparedStatement = connection.prepareStatement("UPDATE cars SET brand = ? WHERE id = ?");
+                preparedStatement.setString(1, brand);
+                preparedStatement.setDouble(2, id);
 
-            int val = preparedStatement.executeUpdate();
-            preparedStatement.close();
-            System.out.println(val + " records were updated");
-        }
-        catch (SQLException e){
-            e.printStackTrace();
+                int val = preparedStatement.executeUpdate();
+                preparedStatement.close();
+                System.out.println(val + " records were updated");
+
+                connection.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
     }
 
@@ -86,16 +87,26 @@ public class Main {
      */
     private static void deleteCarById(long id){
         PreparedStatement preparedStatement = null;
-        try {
-            preparedStatement = connection.prepareStatement("DELETE FROM cars WHERE id = ?;");
-            preparedStatement.setLong(1,id);
 
-            int val = preparedStatement.executeUpdate();
-            preparedStatement.close();
-            System.out.println(val + " records were deleted");
+        Connection connection = getConnection();
+        if(connection != null) {
+            try {
+                preparedStatement = connection.prepareStatement("DELETE FROM cars WHERE id = ?;");
+                preparedStatement.setLong(1, id);
 
-        } catch (SQLException e) {
-            e.printStackTrace();
+                int val = preparedStatement.executeUpdate();
+                preparedStatement.close();
+                System.out.println(val + " records were deleted");
+
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
 
     }
@@ -107,20 +118,29 @@ public class Main {
      * не принимает id т.к. база сгенерирует (auto increment)
      */
     private static void addNewCar(String brand, String model, String bodyStyle, String yearOfManufacture, String color, double price) {
-        try {
-            PreparedStatement preparedStatement = connection.prepareStatement("insert into cars (brand, model, body_style, year_of_manufacture, color, price) values(?, ?, ?, ?, ?, ?)");
-            preparedStatement.setString(1, brand);
-            preparedStatement.setString(2, model);
-            preparedStatement.setString(3, bodyStyle);
-            preparedStatement.setString(4, yearOfManufacture);
-            preparedStatement.setString(5, color);
-            preparedStatement.setDouble(6, price);
+        Connection connection = getConnection();
+        if(connection != null) {
+            try {
+                PreparedStatement preparedStatement = connection.prepareStatement("insert into cars (brand, model, body_style, year_of_manufacture, color, price) values(?, ?, ?, ?, ?, ?)");
+                preparedStatement.setString(1, brand);
+                preparedStatement.setString(2, model);
+                preparedStatement.setString(3, bodyStyle);
+                preparedStatement.setString(4, yearOfManufacture);
+                preparedStatement.setString(5, color);
+                preparedStatement.setDouble(6, price);
 
-            int val = preparedStatement.executeUpdate();
-            preparedStatement.close();
-            System.out.println(val + " records were added to cars");
-        } catch (SQLException e) {
-            e.printStackTrace();
+                int val = preparedStatement.executeUpdate();
+                preparedStatement.close();
+                System.out.println(val + " records were added to cars");
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
 
     }
@@ -132,13 +152,23 @@ public class Main {
      */
     private static ResultSet selectAllFrom(String tableName) {
         ResultSet resultSet = null;
-        String query = "select * from " + tableName;
-        try {
-            Statement statement = connection.createStatement();
-            resultSet = statement.executeQuery(query);
-            statement.close();
-        } catch (SQLException e) {
-            e.printStackTrace();
+
+        Connection connection = getConnection();
+        if(connection != null) {
+            String query = "select * from " + tableName;
+            try {
+                Statement statement = connection.createStatement();
+                resultSet = statement.executeQuery(query);
+                statement.close();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                try {
+                    connection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
         return resultSet;
     }
